move branch_and_bound into the path module
